#3 Gateway configuration
- Gateway module created
- Paths for auth-service module configured
