+++ auth-service/src/main/java/com/karmanov/auth/AuthApp.java
@@ -1,0 +1,11 @@
+package com.karmanov.auth;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AuthApp {
+    public static void main(String[] args) {
+        SpringApplication.run(AuthApp.class,args);
+    }
+}
