--- auth-service/src/main/java/com/karmanov/auth/authservice/AuthProgram.java
@@ -1,11 +1,0 @@
-package com.karmanov.auth.authservice;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AuthProgram {
-    public static void main(String[] args) {
-        SpringApplication.run(AuthProgram.class,args);
-    }
-}
