--- auth-service/src/main/java/com/karmanov/auth/authservice/config/props/ClientProperties.java
@@ -1,34 +1,0 @@
-package com.karmanov.auth.authservice.config.props;
-
-import org.springframework.boot.context.properties.ConfigurationProperties;
-
-@ConfigurationProperties(prefix = "keycloak")
-public class ClientProperties {
-    private String clientId;
-    private String clientSecret;
-    private String redirectUri;
-
-    public String getClientId() {
-        return clientId;
-    }
-
-    public void setClientId(String clientId) {
-        this.clientId = clientId;
-    }
-
-    public String getClientSecret() {
-        return clientSecret;
-    }
-
-    public void setClientSecret(String clientSecret) {
-        this.clientSecret = clientSecret;
-    }
-
-    public String getRedirectUri() {
-        return redirectUri;
-    }
-
-    public void setRedirectUri(String redirectUri) {
-        this.redirectUri = redirectUri;
-    }
-}
