--- auth-service/src/main/java/com/karmanov/auth/authservice/dto/response/UserCreatedRoom.java
@@ -1,6 +1,0 @@
-package com.karmanov.auth.authservice.dto.response;
-
-import java.util.UUID;
-
-public record UserCreatedRoom(UUID id, String name) {
-}
