--- auth-service/src/main/java/com/karmanov/auth/authservice/dto/response/UserProfileInfo.java
@@ -1,7 +1,0 @@
-package com.karmanov.auth.authservice.dto.response;
-
-import java.util.List;
-import java.util.UUID;
-
-public record UserProfileInfo(UUID id, List<UserCreatedRoom> createdRooms, List<UserJoinedRoom> rooms) {
-}
