--- auth-service/src/main/java/com/karmanov/auth/authservice/dto/response/UserRooms.java
@@ -1,6 +1,0 @@
-package com.karmanov.auth.authservice.dto.response;
-
-import java.util.List;
-
-public record UserRooms(List<UserCreatedRoom> createdRooms, List<UserJoinedRoom> joinedRooms) {
-}
