--- auth-service/src/main/java/com/karmanov/auth/authservice/exception/request/RequestFailedException.java
@@ -1,7 +1,0 @@
-package com.karmanov.auth.authservice.exception.request;
-
-public class RequestFailedException extends RuntimeException{
-    public RequestFailedException(String message) {
-        super(message);
-    }
-}
