--- auth-service/src/main/java/com/karmanov/auth/authservice/service/oauth/OauthService.java
@@ -1,5 +1,0 @@
-package com.karmanov.auth.authservice.service.oauth;
-
-public interface OauthService {
-    String getTokens(String code);
-}
