--- auth-service/src/main/java/com/karmanov/auth/authservice/service/user/UserService.java
@@ -1,12 +1,0 @@
-package com.karmanov.auth.authservice.service.user;
-
-import com.karmanov.auth.authservice.dto.response.UserProfileInfo;
-import com.karmanov.auth.authservice.dto.response.UserRooms;
-
-import java.util.UUID;
-
-public interface UserService {
-    UserProfileInfo getOrCreateUser(String keycloakId);
-    UserProfileInfo getUserInfo(UUID userId);
-    UserRooms getUserRooms(UUID userId);
-}
