+++ auth-service/src/main/java/com/karmanov/auth/config/props/ClientProperties.java
@@ -1,0 +1,15 @@
+package com.karmanov.auth.config.props;
+
+import lombok.Getter;
+import lombok.Setter;
+import org.springframework.boot.context.properties.ConfigurationProperties;
+
+@Setter
+@Getter
+@ConfigurationProperties(prefix = "keycloak")
+public class ClientProperties {
+    private String clientId;
+    private String clientSecret;
+    private String redirectUri;
+
+}
