+++ auth-service/src/main/java/com/karmanov/auth/dto/response/ErrorResponse.java
@@ -1,0 +1,21 @@
+package com.karmanov.auth.dto.response;
+
+import lombok.Getter;
+import lombok.Setter;
+
+import java.time.Instant;
+
+@Getter
+@Setter
+public class ErrorResponse {
+    private String message;
+    private String errorCode;
+    private Instant timestamp;
+
+    public ErrorResponse(String message, String errorCode) {
+        this.message = message;
+        this.errorCode = errorCode;
+        this.timestamp = Instant.now();
+    }
+
+}
