+++ auth-service/src/main/java/com/karmanov/auth/dto/response/UserCreatedRoom.java
@@ -1,0 +1,6 @@
+package com.karmanov.auth.dto.response;
+
+import java.util.UUID;
+
+public record UserCreatedRoom(UUID id, String name) {
+}
