+++ auth-service/src/main/java/com/karmanov/auth/dto/response/UserJoinedRoom.java
@@ -1,0 +1,6 @@
+package com.karmanov.auth.dto.response;
+
+import java.util.UUID;
+
+public record UserJoinedRoom(UUID id, String name) {
+}
