+++ auth-service/src/main/java/com/karmanov/auth/dto/response/UserProfileInfo.java
@@ -1,0 +1,7 @@
+package com.karmanov.auth.dto.response;
+
+import java.util.List;
+import java.util.UUID;
+
+public record UserProfileInfo(UUID id, List<UserCreatedRoom> createdRooms, List<UserJoinedRoom> rooms) {
+}
