+++ auth-service/src/main/java/com/karmanov/auth/dto/response/UserRooms.java
@@ -1,0 +1,6 @@
+package com.karmanov.auth.dto.response;
+
+import java.util.List;
+
+public record UserRooms(List<UserCreatedRoom> createdRooms, List<UserJoinedRoom> joinedRooms) {
+}
