+++ auth-service/src/main/java/com/karmanov/auth/exception/request/RequestFailedException.java
@@ -1,0 +1,7 @@
+package com.karmanov.auth.exception.request;
+
+public class RequestFailedException extends RuntimeException{
+    public RequestFailedException(String message) {
+        super(message);
+    }
+}
