+++ auth-service/src/main/java/com/karmanov/auth/service/oauth/OauthService.java
@@ -1,0 +1,5 @@
+package com.karmanov.auth.service.oauth;
+
+public interface OauthService {
+    String getTokens(String code);
+}
