+++ auth-service/src/main/java/com/karmanov/auth/service/request/RequestSender.java
@@ -1,0 +1,8 @@
+package com.karmanov.auth.service.request;
+
+import org.springframework.http.MediaType;
+import org.springframework.util.MultiValueMap;
+
+public interface RequestSender {
+    <T> T sendRequest(MediaType contentType, MultiValueMap<String, String> form, String url, Class<T> responseType);
+}
