+++ build.gradle
@@ -9,11 +9,6 @@ repositories {
     mavenCentral()
 }
 
-dependencies {
-    testImplementation platform('org.junit:junit-bom:5.9.1')
-    testImplementation 'org.junit.jupiter:junit-jupiter'
-}
-
 test {
     useJUnitPlatform()
 }
