+++ gateway-service/src/main/java/com/karmanov/gateway/GatewayApp.java
@@ -1,0 +1,9 @@
+package com.karmanov.gateway;
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+@SpringBootApplication
+public class GatewayApp {
+    public static void main(String[] args) {
+        SpringApplication.run(GatewayApp.class,args);
+    }
+}
