+++ settings.gradle
@@ -1,4 +1,5 @@
 rootProject.name = 'BufferCast'
 include 'clipboard-collector'
 include 'auth-service'
+include 'gateway-service'
